# src/swagger.yaml
openapi: 3.0.0
info:
  title: Google Sheets API
  version: 1.0.0
servers:
  - url: http://localhost:5359

tags:
  - name: Google Sheets API
    description: API for fetching data from Google Sheets

paths:
  /hours-log/andrew/sheet/{dateRange}.json:
    get:
      tags:
        - Google Sheets API
      summary: Fetch data from Google Sheets based on date range
      parameters:
        - in: path
          name: dateRange
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}-\d{4}-\d{2}-\d{2}$'
          required: true
          description: Date range in the format yyyy-mm-dd-yyyy-mm-dd
      responses:
        '200':
          description: A successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  status:
                    type: string
  /hours-log/andrew/sheet.json:
    post:
      tags:
        - Google Sheets API
      summary: Search data from Google Sheets based on original_url
      requestBody:
        required: true
        description: Date range in the format yyyy-mm-dd-yyyy-mm-dd
        content:
          application/json:
            schema:
              type: object
              properties:
                dateRange:
                  type: string
                  pattern: '^\d{4}-\d{2}-\d{2}-\d{4}-\d{2}-\d{2}$'
                  description: Date range in the format yyyy-mm-dd-yyyy-mm-dd
                original_url:
                  type: string
                  description: The URL to search for in the data
              required:
                - dateRange
                - original_url
      responses:
        '200':
          description: A successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  status:
                    type: string

  /hours-log/andrew/sheet/all-data/{dateRange}.json:
    get:
      tags:
        - Google Sheets API
      summary: Fetch all data from the start of the month to today
      parameters:
        - in: path
          name: dateRange
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}-\d{4}-\d{2}-\d{2}$'
          required: true
          description: Date range in the format yyyy-mm-dd-yyyy-mm-dd
      responses:
        '200':
          description: A successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  status:
                    type: string
  /hours-log/andrew/sheet/all-data.json:
    post:
      tags:
        - Google Sheets API
      summary: Search data based on url or original_url
      requestBody:
        required: true
        description: Date range in the format yyyy-mm-dd-yyyy-mm-dd
        content:
          application/json:
            schema:
              type: object
              properties:
                dateRange:
                  type: string
                  pattern: '^\d{4}-\d{2}-\d{2}-\d{4}-\d{2}-\d{2}$'
                  description: Date range in the format yyyy-mm-dd-yyyy-mm-dd
                url:
                  type: string
                  description: The URL to search for in the data
                original_url:
                  type: string
                  description: The original URL to search for in the data
              required:
                - dateRange
                - url
                - original_url
      responses:
        '200':
          description: A successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  status:
                    type: string

  /list-sheets:
    get:
      tags:
        - Google Sheets API
      summary: List all sheets in the Google Sheet
      responses:
        '200':
          description: A successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sheets:
                    type: array
                    items:
                      type: string
                  status:
                    type: string